requirements.txt
# GUI Framework
PyQt5==5.15.9
PyQt6==6.6.0
PyQt6-WebEngine==6.6.0
PySide6==6.6.1

# Web requests and HTTP handling
requests==2.31.0
urllib3==2.0.7
aiohttp==3.9.1
httpx==0.25.2

# XML/HTML parsing
lxml==4.9.3
beautifulsoup4==4.12.2

# JSON and data handling
pyyaml==6.0.1
jsonschema==4.19.2

# Database and ORM
sqlalchemy==2.0.23
sqlite3

# GVM/OpenVAS integration
python-gvm==23.11.0
gvm-tools==23.10.0

# Network scanning
python-nmap==0.7.1
scapy==2.5.0

# Security testing
pycryptodome==3.19.0
cryptography==41.0.7

# Report generation
reportlab==4.0.7
jinja2==3.1.2
matplotlib==3.8.2
plotly==5.17.0
weasyprint==60.2

# Async and threading
asyncio
threading
concurrent.futures

# API testing
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Configuration management
configparser
toml==0.10.2

# Logging and monitoring
colorlog==6.8.0
rich==13.7.0

# Command execution
subprocess32==3.5.4

# File handling
pathlib2==2.3.7

# Date and time
python-dateutil==2.8.2

# Progress bars and UI
tqdm==4.66.1

# Machine Learning and AI
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
transformers==4.33.0

# Data visualization
matplotlib==3.8.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0

# WebSocket and real-time
websockets==11.0.3

# Additional dependencies
OpenGL==3.1.7
PyOpenGL==3.1.7
